requirements.txt
mysql-connector-python
python-dateutil
phonenumbers